initial work on react-router-dom moved pages and App outside /components
