+++ client/package.json
@@ -10,6 +10,7 @@
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-router": "^6.8.1",
+        "react-router-dom": "^6.8.1",
         "react-scripts": "5.0.1",
         "styled-components": "^5.3.6",
         "typescript": "^4.9.5"
